+++ semantic-conventions/src/opentelemetry/semconv/model/semantic_attribute.py
@@ -60,6 +60,7 @@ class SemanticAttribute:
     sampling_relevant: bool
     note: str
     position: List[int]
+    root_namespace: str
     inherited: bool = False
     imported: bool = False
 
@@ -203,6 +204,10 @@ class SemanticAttribute:
             fqn = fqn.strip()
             parsed_brief = TextWithLinks(brief.strip() if brief else "")
             parsed_note = TextWithLinks(note.strip())
+
+            namespaces = fqn.split(".")
+            root_namespace = namespaces[0] if len(namespaces) > 1 else ""
+
             attr = SemanticAttribute(
                 fqn=fqn,
                 attr_id=attr_id,
@@ -218,6 +223,7 @@ class SemanticAttribute:
                 sampling_relevant=sampling_relevant,
                 note=parsed_note,
                 position=position,
+                root_namespace=root_namespace,
             )
             if attr.fqn in attributes:
                 position = position_data[list(attribute)[0]]
@@ -337,7 +343,6 @@ class SemanticAttribute:
         elif self == other:
             return True
         return False
-
 
 class AttributeType: